+++ app/src/main/res/xml/daily_messages_widget_info.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:configure="com.sapuseven.untis.widgets.DailyMessagesWidgetConfigureActivity"
+    android:configure="com.sapuseven.untis.activities.DailyMessagesWidgetConfigureActivity"
     android:initialKeyguardLayout="@layout/daily_messages_widget"
     android:initialLayout="@layout/daily_messages_widget"
     android:minWidth="180dp"
